+++ tests/deprecated_api/test_remove_1-7.py
@@ -61,11 +61,6 @@ def test_v1_7_0_on_interrupt(tmpdir):
         trainer.fit(model)
 
 
-def test_v1_7_0_process_position_trainer_constructor(tmpdir):
-    with pytest.deprecated_call(match=r"Setting `Trainer\(process_position=5\)` is deprecated in v1.5"):
-        _ = Trainer(process_position=5)
-
-
 def test_v1_7_0_flush_logs_every_n_steps_trainer_constructor(tmpdir):
     with pytest.deprecated_call(match=r"Setting `Trainer\(flush_logs_every_n_steps=10\)` is deprecated in v1.5"):
         _ = Trainer(flush_logs_every_n_steps=10)